Worked on the front-end development of website, attempted the login function
