+++ src/App.css
@@ -9,12 +9,11 @@ body, #root {
 
 .home{
   background: linear-gradient(180deg, #12122e 0%, #272764 100%);
-
   line-height: 38x;
   font-weight: 600;
   font-size: 1.125rem;
   width: 100%;
-  height: 120vh;
+  height: 100vh;
   color: #ffff;
   display: flex;
   flex-direction: column;
