+++ src/index.css
@@ -69,6 +69,7 @@ span.highlighted {
 }
 
 
+
 .title-section-container .title-info-container {
   height: 80%;
   display: flex;
